+++ com.unity.render-pipelines.lightweight/Editor/LWRPAdditionalCameraDataEditor.cs
@@ -1,6 +1,6 @@
-using UnityEngine.Experimental.Rendering.LightweightPipeline;
+using UnityEngine.Rendering.LWRP;
 
-namespace UnityEditor.Experimental.Rendering.LightweightPipeline
+namespace UnityEditor.Rendering.LWRP
 {
     [CanEditMultipleObjects]
     // Disable the GUI for additional camera data