+++ com.unity.render-pipelines.lightweight/Runtime/Passes/SceneViewDepthCopy.cs
@@ -1,7 +1,8 @@
 using System;
 using UnityEngine.Rendering;
+using UnityEngine.Rendering.LWRP;
 
-namespace UnityEngine.Experimental.Rendering.LightweightPipeline
+namespace UnityEngine.Experimental.Rendering.LWRP
 {
     internal class SceneViewDepthCopyPass : ScriptableRenderPass
     {